refactor : make child code text field file
