+++ alimo/Projects/App/Sources/UI/Feature/Onboarding/ParentJoin/ParentJoinFirstView.swift
@@ -121,83 +121,3 @@ struct ParentJoinFirstView: View {
         }
     }
 }
-
-struct CharacterField: View {
-    @State var character: String = ""
-    @FocusState var focused: Int?
-    @Binding var characters: [Int : String]
-    
-    var index: Int
-    var onChange: ((_ index: Int, _ char: String) -> Void)
-    
-    var body: some View {
-        TextField(text: $character) {
-            Text("")
-                .font(.body)
-        }
-        .lineLimit(1)
-        .multilineTextAlignment(.center)
-        .background(
-            RoundedRectangle(cornerRadius: 8)
-                .foregroundStyle(.white)
-                .frame(height: 52)
-                .clipShape(RoundedRectangle(cornerRadius: Size.large.rawValue))
-                .overlay {
-                    if focused == index {
-                        RoundedRectangle(cornerRadius: 8)
-                            .stroke(Color.gray500, lineWidth: 2)
-                    } else {
-                        if characters[index] != nil && characters[index] != "" {
-                            RoundedRectangle(cornerRadius: 8)
-                                .stroke(Color.main500, lineWidth: 2)
-                        } else {
-                            RoundedRectangle(cornerRadius: 8)
-                                .stroke(Color.gray300, lineWidth: 2)
-                        }
-                    }
-                }
-        )
-        .frame(maxWidth: .infinity, alignment: .center)
-        .focused($focused, equals: index)
-        .onChange(of: character) { newValue in
-            onChange(index, character)
-            
-            if newValue.count > 1 {
-                character = String(newValue[newValue.startIndex..<newValue.index(newValue.startIndex, offsetBy: 1)])
-            }
-        }
-        .padding(.horizontal, 3)
-        
-    }
-}
-
-struct SeparatedTextField: View {
-    var length: Int // 입력받은 글자 수
-    @FocusState var focused: Int?
-    // index별로 글자를 저장할 dictionary
-    @State var characters: [Int : String] = [:]
-    
-    @Binding var string: String
-    
-    var body: some View {
-        HStack {
-            ForEach(0..<length) { i in
-                CharacterField(focused: _focused, characters: $characters, index: i) { i, c in
-                    focused = c.isEmpty ? i - 1 : i + 1
-                    characters[i] = c
-                    string = getString()
-                }
-            }
-        }.padding([.vertical], 16)
-    }
-    
-    func getString() -> String {
-        var str = ""
-        for i in 0..<length {
-            if let c = characters[i] {
-                str += c
-            }
-        }
-        return str
-    }
-}
